+++ apps/monitor-web/src/components/common/index.ts
@@ -12,4 +12,4 @@ export { default as Toast } from "./Toast";
 export { default as Image } from "./Image";
 export { default as BasicButton } from "./Buttons/BasicButton";
 export { default as IconButton } from "./Buttons/IconButton";
-export { default as Modal } from "./Modal";
+export { default as ModalLayout } from "./ModalLayout";
